autostream-backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
langchain-huggingface
langchain-text-splitters
pydantic
python-dotenv
faiss-cpu
sentence-transformers
tiktoken
